Refactor ATS score calculation in ResumeBuilder for improved scoring logic and update PDF export styles in ResumePreview for better formatting. Change ats_score column type to FLOAT in database migration for more accurate score representation.

+++ supabase/migrations/20240601000001_create_users_and_resumes_tables.sql
@@ -18,7 +18,7 @@ CREATE TABLE IF NOT EXISTS public.resumes (
   skills JSONB NOT NULL,
   achievements JSONB NOT NULL,
   selected_template TEXT NOT NULL,
-  ats_score INTEGER,
+  ats_score FLOAT,
   created_at TIMESTAMP WITH TIME ZONE DEFAULT now() NOT NULL,
   updated_at TIMESTAMP WITH TIME ZONE DEFAULT now() NOT NULL
 );
